Use 0 compression when creating the target_files package
The zip file under $OUT/obj/PACKAGING/target_files_intermediate
is used for final package creation, and it's being
unzipped multiple times by releasetools.

(Not to mention that it also takes an awful lot of
time just to zip it initially.)

Using 0 compression on it reduces the build time by
several minutes.

The final zip package is untouched by this change.

Change-Id: Ic38b470f97f1e0ebb80da9d4f7750e4ccf53e05c